Refactoring: extracted web/homepage.go
