+++ cmd/web.go
@@ -1,34 +1,27 @@
 package cmd
 
 import (
-  "bytes"
-  "encoding/json"
   "fmt"
   "github.com/oaktown/calliope/misc"
-  "github.com/oaktown/calliope/report"
-  "github.com/oaktown/calliope/store"
-  "github.com/olivere/elastic"
+  "github.com/oaktown/calliope/web"
   "github.com/spf13/cobra"
-  "html/template"
   "log"
   "net/http"
-  "time"
 )
 
-var port, label, url string
-var size int
+var port string
 var allMessages bool
-
+var options = web.Options{}
 
 func init() {
   rootCmd.AddCommand(webCmd)
   webCmd.Flags().StringVarP(&port, "port", "p", "8080", "Port to run web server on.")
-  webCmd.Flags().StringVarP(&label, "label", "l", "",
+  webCmd.Flags().StringVarP(&options.Label, "label", "l", "",
     "Report for emails with this label (required).")
   webCmd.MarkFlagRequired("label")
   webCmd.Flags().BoolVarP(&allMessages, "all-messages", "A", false, "By default, we only query for starred messages. With this flag, we get all messages for the label whether they are starred or not.")
-  webCmd.Flags().IntVarP(&size, "size", "s", 1000, "The max number of results to return from a search. Defaults to 1000.")
-  webCmd.Flags().StringVarP(&url, "url", "U", "https://mail.google.com/mail/", "Url for gmail (useful if you are logged into multiple accounts).")
+  webCmd.Flags().IntVarP(&options.Size, "size", "s", 1000, "The max number of results to return from a search. Defaults to 1000.")
+  webCmd.Flags().StringVarP(&options.InboxUrl, "url", "U", "https://mail.google.com/mail/", "Url for gmail (useful if you are logged into multiple accounts).")
 }
 
 var webCmd = &cobra.Command{
@@ -36,22 +29,12 @@ var webCmd = &cobra.Command{
   Short: "Start web server",
   Long:  `Starts up the Calliope web app.`,
   Run: func(cmd *cobra.Command, args []string) {
-    web()
+    options.Starred = !allMessages
+    startServer(options)
   },
 }
 
-type Data struct {
-  Title       string
-  Size        int
-  Query       string
-  Report      template.HTML
-  Earliest    time.Time
-  Latest      time.Time
-  TotalEmails int64
-}
-
-func web() {
-  starred := !allMessages
+func startServer(opt web.Options) {
   client := misc.GetStoreClient()
   initialQuery := true
 
@@ -61,53 +44,17 @@ func web() {
 
   http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
     var query string
+
     if initialQuery {
-      query = GetInitialQueryString(client, starred)
+      query = web.QueryStringFromLabel(client, opt)
       initialQuery = false
     } else {
       query = r.FormValue("query")
     }
 
-    ShowHomePage(client, query, w)
+    web.ShowHomePage(client, query, w, opt)
   })
 
   fmt.Printf("Starting web server: http://localhost:%s/\n\n", port)
   log.Fatal(http.ListenAndServe(":"+port, nil))
-}
-
-func ShowHomePage(client *store.Service, query string, w http.ResponseWriter) {
-  stats, _ := client.GetStats()
-  reportHtml := template.HTML(GetReportHtml(client, query))
-  data := Data{
-    Title:       "Calliope Email Report",
-    Size:        size,
-    Query:       query,
-    Report:      reportHtml,
-    Earliest:    stats.Earliest,
-    Latest:      stats.Latest,
-    TotalEmails: stats.Total,
-  }
-
-  t := template.Must(template.ParseFiles("templates/layout.html", "templates/web-ui.html"))
-
-  if err := t.ExecuteTemplate(w, "layout", data); err != nil {
-    log.Println("Error occurred while executing template: ", err)
-  }
-}
-
-func GetReportHtml(client *store.Service, query string) string {
-  var req *elastic.SearchService
-  req = client.GetRawQuery(query, size)
-  var reportBuffer bytes.Buffer
-  report.Run(client, &reportBuffer, req, inboxUrl)
-  return reportBuffer.String()
-}
-
-func GetInitialQueryString(client *store.Service, starred bool) string {
-  boolQuery, _ := client.GetQueryFromLabel(label, starred, size)
-  source, _ := boolQuery.Source()
-  q, _ := json.MarshalIndent(source, "  ", "  ")
-  query = fmt.Sprintf("{\n  \"query\": %s\n}", q)
-  fmt.Println("Query using label: ", query)
-  return query
 }
